refactor: Move test_SKOpt to own file
